Implement console application for taxi service aggregation
Implements CommandLineArguments and ConsoleApplication.

CommandLineArguments is a data structure that validates command line
arguments that are inputted as a String[] in the main method. Also,
CLA does not throw exceptions when user inputted arguments are invalid.
Instead, CLA uses a boolean member variable to notify the program if the
data is invalid. This allows the program to exit more gracefully than
scaring away user with stacktraces and full class names.

ConsoleApplication is the bread and butter of the program that
aggregates the user's query and prints it to standard output.