UNSTABLE: Add simplified version of console_app
Adds a simplified version of the console_app that only takes data from
Dave's Taxis and set the capacity limit to 1.

Though I have not spotted any problems so far, this code was quick
solution by reducing down a copy of the console_app and aggregator
sub-projects, and has nowhere near the same code coverage. This will
most likely not be merged to the master branch, but instead kept on an
unstable branch.