Update documentation to reflect new subproject dave_console_app
Removes old information about the unstable branch, which will be will
not be included in the submission (only for curiosity's sake).

Adds information on how to run new subproject. Updates some of the some
of the old documentation to include the changes done to running the
console_app project after the dave_console_app commit.

Adds examples of how to run Part 2. Clean and tweak the document a bit
to structure document closer to sample submission and make the document
easier to follow.

Added some more information about the project itself.